sensor-simulator/requirements.txt
pyyaml
bacpypes3
pymodbus==3.8.0
